tests/wpt/web-platform-tests/css/css-masking/parsing/clip-rule-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing clip-rule with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-clip-rule">
<meta name="assert" content="clip-rule supports the full grammar 'nonzero | evenodd'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("clip-rule", "nonzero");
test_valid_value("clip-rule", "evenodd");
</script>
</body>
</html>
